+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🚩 Deprecations 🚩
 
 - Deprecate `p[metric|log|trace]otlp.RegiserServer` in favor of `p[metric|log|trace]otlp.RegiserGRPCServer` (#6180)
+- Deprecate `pcommon.Map.PutString` in favor of `pcommon.Map.PutStr` (#6210)
 
 ### 💡 Enhancements 💡